requirements.txt
PyYAML
pydantic
prometheus-client
paramiko
